cpp: allow implicit conversion from boolean to bool
This allows the use of isl::boolean values as first operand of
__builtin_assume(long, long) without the need for explicit casts. This is
useful as OSX uses __builtin_assume in its implementation of assert() to
indicate to the compiler that the condition that is verified is true with high
probability. Without allowing implicit conversions to bool, the use
of isl::boolean without an explicit cast to bool results in
a compilation failure on OSX. As explicit casts seem too verbose to be used in
each assert call, this change allows implicit conversions from boolean
to bool.

Originally, implicit conversions from boolean to bool were disallowed to start
off with a minimal C++ interface that is as explicit as possible.  This
minimal interface was understood to be widened in case practical use
cases motivate an extended interface.  Extending the interface to avoid
many explicit casts in common situtations is exactly such a practical
use case.

Also update the tests. Incorrectly, the tests already assumed that
'assert(b_true)' would test implicit conversions. This was not the case
and new test cases have been added that test implicit conversion to both
bool and long.

Suggested-by: Siddharth Bath <siddharth.bhat@research.iiit.ac.in>
Signed-off-by: Tobias Grosser <tobias@grosser.es>