interface/cpp.cc: print opening brace of function together with closing brace
In particular, the opening brace was printed in
cpp_generator::print_method_header, while the closing brace was
printed in cpp_generator::print_method_impl.
This is confusing.
Print both in cpp_generator::print_method_impl.

As a side-effect, the opening brace is printed on the next line,
which is more in line with the isl coding style (from which
the C++ bindings are known to deviate).

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>