move bindings generation to the top level
Generating the bindings at the top level makes it easier
to choose where the bindings will be stored.
This is especially useful for the C++ bindings (with exceptions)
that will be added later and that will be generated directly
under include/ (and not somewhere inside interface/).
The location of the currently generated bindings is not changed
by this commit.

An additionally advantage is that the test executables
now properly depend on the libisl.la library.

As a side effect, "make dist" now fails if no isl-noexceptions.h
can be generated, but this may also be considered an advantage.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>