add isl_*_list_get as an alternative name for isl_*_list_get_*
The type of the element is implicit in the type of the list,
so it does not need to be specified explicitly in the name of the function.
Moreover, several other isl_*_list functions already omit this element type
in their names.

This change helps to simplify the tests in the following get_list patch
and is also tested there.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>
Signed-off-by: Tobias Grosser <tobias@grosser.es>