Ensure isl_ast_graft_free returns a pointer to isl_ast_graft
While the currently returned void pointer is automatically converted
to "isl_reordering *" where needed, it is both clearer and more in
line with other isl functions to return a pointer to the type that is
freed. This also allows to catch bugs where the return type is
implicitly converted to a different type.

Signed-off-by: Tobias Grosser <tobias@grosser.es>