drop namespace qualifier in generated C++ bindings
In the checked C++ bindings, the types live in the namespace
"isl::checked" rather than in the namespace "isl".
It's currently possible to refer to them as living in the "isl" namespace
because the nested "checked" namespace is inlined.
However, the next commit will remove the inlining.

Remove the namespace qualification such that the types
are simply looked up in the current namespace, whatever it is.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>