generate conversion functions between default and checked C++ bindings
A user who cannot (or prefers not to) use exceptions may want
to use the default bindings by default (which will abort
on any error) and the checked bindings in selected places
where it is important to recover from errors.
The functions generated by this commit allow objects
to be converted from one set of bindings to another
at the point where the users wants to make the switch.

Note that if exceptions are available, then the user can simply
use the default bindings and catch exceptions thrown in the selected pieces,
while ignoring them in other parts.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>