isl_test_cpp-checked: consistently use is_true() to test isl::checked::boolean
In some cases, the conversion to bool was used instead.
This conversion operator currently only complains
if the isl::checked::boolean is in an error state,
but it will be changed to complain if the error state
has not been checked yet.

Explicitly call is_true() instead.
For the generic tests, which are reused for the default bindings,
hide the call behind an IS_TRUE macro such that the tests can
continue to be reused.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>