checked C++ bindings: abort on unchecked boolean error when converting to bool
The current bindings only complain when an actual error occurs,
but this does not provide any kind of safety, since missing
error checks are only detected when an actual error condition occurs.
Improve on this slightly by also complaining when
an isl::checked:boolean is converted to bool without a prior
check for errors.  If is_error() is called, then there clearly
has been such a check.  A call to is_false() or isl_true()
is also taken a check for errors since these function returning
true means that no error has occurred.  The assumption here is
that the user would tend to write is_false() instead of !is_true().

Ideally, the check for errors should be enforced by the compiler,
but this is left for future work.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>