checked C++ bindings: abort on unchecked stat error
That is, require a call to is_ok()/is_error() on every
isl::checked::stat object.
This is similar to the way isl::checked::boolean is handled,
but slightly stricter because calling one of these methods
is always sufficient for determining if an error state has occurred.
When the isl_stat value is released, it is assumed to be sent
to isl, which will also check the error state.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>