make cpp_generator::type2cpp(isl_class) public (and static)
It is not clear why this method, introduced in isl-0.18-598-g81c38ef4c
(cpp: generate C++ wrapper classes, Sat Apr 8 05:12:50 2017 +0200),
wasn't made static in the first place.
Turn it into a public method such that it can be reused
by the generator for conversions between the default and
the checked C++ bindings of type based subclasses.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>