vertices: rename isl_external_vertex to isl_vertex
This ensures that the external type 'isl_vertex' actually is an
appreviation for 'struct isl_vertex'. Before this change set
'isl_vertex' refererred to 'struct isl_external_vertex' and there was an
internal structure 'struct isl_vertex', which appered to be the
implementation of 'isl_vertex' but was not.

To make the name 'struct isl_vertex' available, the internal
implementation is now called 'struct isl_internal_vertex'.

Signed-off-by: Tobias Grosser <tobias@grosser.es>